+++ src/main/java/com/sven/machine/learning/mnist/MnistData.java
@@ -3,17 +3,17 @@ package com.sven.machine.learning.mnist;
 public class MnistData {
 
 	private int label;
-	private int[][] imageByte;
+	private double[][] imageByte;
 	public int getLabel() {
 		return label;
 	}
 	public void setLabel(int label) {
 		this.label = label;
 	}
-	public int[][] getImageByte() {
+	public double[][] getImageByte() {
 		return imageByte;
 	}
-	public void setImageByte(int[][] imageByte) {
+	public void setImageByte(double[][] imageByte) {
 		this.imageByte = imageByte;
 	}
 	@Override
